+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-
 ## [2022-05-31] v0.9.6
 
 - 🐞 [issues/4](https://github.com/vite-plugin/vite-plugin-dynamic-import/issues/4)
 - 🌱 Support `import(node_modules/xxx)`
+
+## [2022-06-03] v0.9.7
+
+- refactor: remove vite-plugin-utils
+- refactor: merge alias.ts into resolve.ts
+- refactor: use resolve only
+- refactor: migration alias to resolve